Clean up SslAesTransport and move SmartCamera
