+++ kasa/experimental/__init__.py
@@ -1,0 +1,7 @@
+"""Package for experimental."""
+
+from .smartcamera import SmartCamera
+
+__all__ = [
+    "SmartCamera",
+]
