+++ kasa/smart/__init__.py
@@ -1,7 +1,6 @@
 """Package for supporting tapo-branded and newer kasa devices."""
 
-from .smartcamera import SmartCamera
 from .smartchilddevice import SmartChildDevice
 from .smartdevice import SmartDevice
 
-__all__ = ["SmartDevice", "SmartChildDevice", "SmartCamera"]
+__all__ = ["SmartDevice", "SmartChildDevice"]
